Export AmqpAnnotatedMessage from core-amqp
