+++ sdk/core/core-amqp/src/index.ts
@@ -62,4 +62,5 @@ export {
   isNode,
   WebSocketOptions
 } from "./util/utils";
+export { AmqpAnnotatedMessage } from "./amqpAnnotatedMessage";
 export { logger } from "./log";
